Move calculate_qc_metrics out of simple.py
