+++ scanpy/api/pp.py
@@ -1,7 +1,7 @@
 from ..preprocessing.recipes import recipe_zheng17, recipe_weinreb17, recipe_seurat
 from ..preprocessing.simple import filter_cells, filter_genes, filter_genes_dispersion
 from ..preprocessing.simple import log1p, sqrt, pca, normalize_per_cell, regress_out, scale, subsample, downsample_counts
-from ..preprocessing.simple import calculate_qc_metrics
+from ..preprocessing.qc import calculate_qc_metrics
 from ..preprocessing.mnn_correct import mnn_correct
 from ..preprocessing.dca import dca
 from ..preprocessing.magic import magic
