+++ scanpy/preprocessing/simple.py
@@ -5,12 +5,11 @@ Compositions of these functions are found in sc.preprocess.recipes.
 
 import scipy as sp
 import warnings
-from scipy.sparse import issparse, isspmatrix_coo, csr_matrix
+from scipy.sparse import issparse, csr_matrix
 from sklearn.utils import sparsefuncs
 import pandas as pd
 from pandas.api.types import is_categorical_dtype
 from anndata import AnnData
-from .qc import top_segment_proportions
 from .. import settings as sett
 from .. import logging as logg
 from ..utils import sanitize_anndata
@@ -1084,106 +1083,6 @@ def downsample_counts(adata, target_counts=20000, random_state=0, copy=False):
             adata.X[icell] = cell_profile
     logg.msg('finished', t=True)
     return adata if copy else None
-
-
-def calculate_qc_metrics(adata, exprs_values="counts", feature_controls=(), 
-                         percent_top=(50, 100, 200, 500), inplace=False):
-    """
-    Calculate quality control metrics.
-
-    Calculates a number of qc metrics for an AnnData object, largely based on
-    `calculateQCMetrics` from scater [McCarthy17]. Currently is most efficient
-    on a sparse CSR or dense matrix.
-
-    Parameters
-    ----------
-    adata : :class:`~anndata.AnnData`
-        Annotated data matrix.
-    exprs_values : `str`, optional (default: "counts")
-        Name of kind of values in X.
-    feature_controls : `Container`, optional (default: `()`)
-        Keys for boolean columns of `.var` which identify feature controls,
-        e.g. "ERCC" or "mito".
-    percent_top : `Collection[int]`, optional (default: (50, 100, 200, 500))
-        Which proportions of top genes to cover. If empty or `None` don't
-        calculate.
-    inplace : bool, optional (default: `False`)
-        Whether to place calculated metrics in `.obs` and `.var`
-
-    Returns
-    -------
-    Union[NoneType, Tuple[pd.DataFrame, pd.DataFrame]]
-        Depending on `inplace` returns calculated metrics (`pd.DataFrame`) or
-        updates `adata`'s `obs` and `var`.
-
-        Observation level metrics include:
-
-        * `total_features_by_{exprs_values}`
-        * `total_{expr_values}`
-        * `pct_{expr_values}_in_top_{n}_features` - for `n` in `percent_top`
-        * `total_{exprs_values}_{feature_control}` - for each `feature_control`
-        * `pct_{exprs_values}_{feature_control}` - for each `feature_control`
-
-        Variable level metrics include:
-
-        * `total_{expr_values}`
-        * `mean_{expr_values}`
-        * `n_cells_by_{expr_values}`
-        * `pct_dropout_by_{expr_values}`
-    """
-    if isspmatrix_coo(adata.X):
-        X = csr_matrix(adata.X)  # COO not subscriptable
-    else:
-        X = adata.X
-    obs_metrics = pd.DataFrame(index=adata.obs_names)
-    var_metrics = pd.DataFrame(index=adata.var_names)
-    # Calculate obs metrics
-    obs_metrics["total_features_by_{exprs_values}"] = (
-        X != 0).sum(axis=1)
-    obs_metrics["log1p_total_features_by_{exprs_values}"] = np.log1p(
-        obs_metrics["total_features_by_{exprs_values}"])
-    obs_metrics["total_{exprs_values}"] = X.sum(axis=1)
-    obs_metrics["log1p_total_{exprs_values}"] = np.log1p(
-        obs_metrics["total_{exprs_values}"])
-    proportions = top_segment_proportions(X, percent_top)
-    # Since there are local loop variables, formatting must occur in their scope
-    # Probably worth looking into a python3.5 compatable way to make this better
-    for i, n in enumerate(percent_top):
-        obs_metrics["pct_{exprs_values}_in_top_{n}_features".format(**locals())] = \
-            proportions[:, i] * 100
-    for feature_control in feature_controls:
-        obs_metrics["total_{exprs_values}_{feature_control}".format(**locals())] = \
-            X[:, adata.var[feature_control].values].sum(axis=1)
-        obs_metrics["log1p_total_{exprs_values}_{feature_control}".format(**locals())] = \
-            np.log1p(
-                obs_metrics["total_{exprs_values}_{feature_control}".format(**locals())])
-        # "total_{exprs_values}" not formatted yet
-        obs_metrics["pct_{exprs_values}_{feature_control}".format(**locals())] = \
-            obs_metrics["total_{exprs_values}"] / \
-            obs_metrics["total_{exprs_values}_{feature_control}".format(
-                **locals())]
-    # Calculate var metrics
-    var_metrics["mean_{exprs_values}"] = np.ravel(X.mean(axis=0))
-    var_metrics["log1p_mean_{exprs_values}"] = np.log1p(
-        var_metrics["mean_{exprs_values}"])
-    var_metrics["n_cells_by_{exprs_values}"] = np.ravel((X != 0).sum(axis=0))
-    var_metrics["pct_dropout_by_{exprs_values}"] = \
-        (1 - var_metrics["n_cells_by_{exprs_values}"] / X.shape[0]) * 100
-    var_metrics["total_{exprs_values}"] = np.ravel(X.sum(axis=0))
-    var_metrics["log1p_total_{exprs_values}"] = np.log1p(
-        var_metrics["total_{exprs_values}"])
-    # Format strings
-    for df in obs_metrics, var_metrics:
-        new_colnames = []
-        for col in df.columns:
-            new_colnames.append(col.format(**locals()))
-        df.columns = new_colnames
-    # Return
-    if inplace:
-        adata.obs = adata.obs.join(obs_metrics)
-        adata.var = adata.var.join(var_metrics)
-    else:
-        return obs_metrics, var_metrics
 
 
 def zscore_deprecated(X):
